Forum post & comments public api endpoints (#3527)
* Setup ceramic staging instance (#3503)

* Setup ceramic staging instance

* Add missing config

* Perform pre prod cleanup

* fix file

* Use state from main

* Merge

* Fix state

* Delete focalboard

* Readd focalboard

* WIP caching

* Fix caching approach

* Fix

* Fix apollo client handler to avoid using Redis if not ready

* Fix set cache bug

* Cleanup errors

* Trigger CI

* fix docker

* Add ceramic build step and use base node to launch process

* Clenup files

* Use new environment

* Fix ceramic link

* Add prod ceramic docker compose

* Use real user credentials

* Github action: automated Core pk upgrade to 0.25.0

* Set credential logo for space (#3508)

* Set credential logo for space

* Add iconUrl for internal attestations based on space credential logo

* refactor

* OTP wallet login (#3497)

* OTP wallet login

* Code cleanup

* Remove setLoaded from test render

* Added modal for wallet connect

* Show message only if user ha authenticated

* Reverted back reload logic.

* Fix displaying users in details view (#3518)

* Fix ceramic (#3519)

* Show proposal credential placeholder in author property (#3520)

* Show proposal credential placeholder

* Add log to catch failure in issuing credentials to proposal

* add ability to add a workflow or reward template from proposal page (#3510)

* add ability to add a workflow or reward template from proposal page

* .

* disable add new for rewards

* update snapshots

* fix class

* remove prop

* remove unused prop

* Set primary wallet & use during credential issue process (#3509)

* Set primary wallet & use during credential issue process

* Resolved PR Comments

* Remove ceramic node from cron

* Fixed image upload issue with credential logo

---------


* Github action: automated Core pk upgrade to 0.26.0

* assign a uuid when creating a new reward (#3524)

* bump datadog trace (#3516)

* upgrade lodash

* bump dd-trace lib

* fix convert to propoosal (#3525)

* fix snapshot voting (#3526)

* fix snapshot voting

* .

* layout

* Include author with forum posts

* Added endpoint to create forum post

* Added post update endpoint

* Full support for upvoting/downvoting post

* Updated endpoints to create and get post comments

* Fixed post comment edit and delete endpoints

* Finished post comment upvote endpoint

* Update Ceramic model and drop feature flags (#3522)

* Update model and drop feature flags

* Handle empty state for credentials

* Revert "Update Ceramic model and drop feature flags (#3522)" (#3528)

This reverts commit c7f15ab5e9772a15824b4ff45fcdb72810d0d637.

---------

+++ pages/api/v1/forum/posts/[postId]/comments/[commentId]/upvote.ts
@@ -1,7 +1,9 @@
+import { UnauthorisedActionError } from '@charmverse/core/errors';
 import { prisma } from '@charmverse/core/prisma-client';
 import type { NextApiRequest, NextApiResponse } from 'next';
 
 import { requireKeys } from 'lib/middleware';
+import { generateMarkdown } from 'lib/prosemirror/plugins/markdown/generateMarkdown';
 import { superApiHandler } from 'lib/public-api/handler';
 import { withSessionRoute } from 'lib/session/withSession';
 
@@ -9,46 +11,35 @@ import type { PublicApiPostComment } from '../index';
 
 const handler = superApiHandler();
 
-handler.post(requireKeys(['userId'], 'body'), voteOnComment);
+handler.post(requireKeys(['userId', 'upvoted'], 'body'), upvoteOnComment);
 
 /**
  * @swagger
  * /forum/posts/{postId}/comments/{commentId}/vote:
  *   post:
- *     summary: Up/downvote a proposal comment
- *     description: Adds a vote for a proposal comment by a specific user
+ *     summary: Up/downvote a post comment
+ *     description: Adds a vote for a post comment by a specific user
  *     tags:
  *      - 'Partner API'
  *     parameters:
- *       - name: proposalIdOrPath
- *         in: path
- *         required: true
- *         type: string
- *         description: ID or page path of the related proposal
  *       - name: postId
  *         in: path
  *         required: true
+ *         schema:
+ *           type: string
+ *         description: ID of the related post
+ *       - name: commentId
+ *         in: params
+ *         required: true
  *         type: string
- *         description: ID of the comment to create a vote for
+ *         description: ID of the comment to update
  *     requestBody:
  *       required: true
  *       content:
  *         application/json:
  *           schema:
  *             type: object
- *             properties:
- *               userId:
- *                 type: string
- *                 description: User ID of the user who is performing the upvote / downvote
- *                 example: "69a54a56-50d6-4f7b-b350-2d9c312f81f3"
- *               upvoted:
- *                 type: boolean
- *                 nullable: true
- *                 description: true for an upvote, false for a downvote, null to delete the user's upvote / downvote
- *                 example: true
- *             required:
- *               - userId
- *               - upvoted
+ *             $ref: '#/components/schemas/UpvoteInput'
  *     responses:
  *       200:
  *         description: Comment where the vote was made with refreshed vote count
@@ -56,10 +47,10 @@ handler.post(requireKeys(['userId'], 'body'), voteOnComment);
  *            application/json:
  *              schema:
  *                type: object
- *                $ref: '#/components/schemas/ProposalComment'
+ *                $ref: '#/components/schemas/ForumPostComment'
  *
  */
-async function voteOnComment(req: NextApiRequest, res: NextApiResponse<PublicApiPostComment>) {
+async function upvoteOnComment(req: NextApiRequest, res: NextApiResponse<PublicApiPostComment>) {
   // This should never be undefined, but adding this safeguard for future proofing
   const userId = req.body.userId as string;
   const postId = req.query.postId as string;
@@ -76,6 +67,19 @@ async function voteOnComment(req: NextApiRequest, res: NextApiResponse<PublicApi
       id: true
     }
   });
+
+  const spaceRole = await prisma.spaceRole.findFirst({
+    where: {
+      spaceId: {
+        in: req.spaceIdRange
+      },
+      userId
+    }
+  });
+
+  if (!spaceRole) {
+    throw new UnauthorisedActionError('User does not have access to this space');
+  }
 
   if (req.body.upvoted === null) {
     await prisma.postCommentUpDownVote.delete({
@@ -106,7 +110,48 @@ async function voteOnComment(req: NextApiRequest, res: NextApiResponse<PublicApi
     });
   }
 
-  return res.status(200).json({});
+  const comment = await prisma.postComment.findUniqueOrThrow({
+    where: {
+      id: commentId
+    },
+    include: {
+      votes: {
+        select: {
+          upvoted: true
+        }
+      }
+    }
+  });
+
+  const parsedContent = await generateMarkdown({
+    content: comment.content
+  });
+
+  const { upvotes, downvotes } = comment.votes.reduce(
+    (acc, val) => {
+      if (val.upvoted) {
+        acc.upvotes += 1;
+      } else if (val.upvoted === false) {
+        acc.downvotes += 1;
+      }
+      return acc;
+    },
+    { upvotes: 0, downvotes: 0 }
+  );
+
+  return res.status(200).json({
+    id: comment.id,
+    createdBy: comment.createdBy,
+    createdAt: comment.createdAt.toISOString(),
+    parentId: comment.parentId,
+    content: {
+      markdown: parsedContent,
+      text: comment.contentText
+    },
+    upvotes,
+    downvotes,
+    children: []
+  });
 }
 
 export default withSessionRoute(handler);
